Rename JobComponentHandlerInterface -> JobComponentPreparationFactory (#1341)
