+++ config/services_test.yaml
@@ -44,7 +44,7 @@ services:
         arguments:
             $httpClient: '@app.tests.integration.http.client'
 
-    App\Services\WorkerJobFactory:
+    App\Services\WorkerJobComponentFactory:
         public: true
 
     App\Services\PreparationStateFactory:
